requirements.txt
# Core Dependencies
tensorflow>=2.15.0
keras>=3.0.0
numpy>=1.24.0
opencv-python>=4.8.0
mediapipe>=0.10.9

# Web Interface
streamlit>=1.28.0

# Audio
pyttsx3>=2.90

# Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.4.0

# Utilities
tqdm>=4.66.0
pillow>=10.0.0

# Optional: GPU Support
# tensorflow-gpu>=2.15.0
